Extracted raml expansion methods to their own module Changed travis tests to test Node.JS 4,5,6,7,8
