+++ .travis.yml
@@ -8,9 +8,10 @@ notifications:
     on_failure: change
 
 node_js:
-  - "0.12"
   - "4"
-  - "node"
-  - "iojs"
+  - "5"
+  - "6"
+  - "7"
+  - "8"
 
 after_script: "npm install coveralls@2 && cat ./coverage/lcov.info | coveralls"
